Rename test files to be consistent with other file names.
